Performance: if ignoreOriginal is set, do not save the original object serialized in annotations
Premise: wrangler's apply mechanism uses a three way patch calculated from the original object
(latest version that was ever applied), the current object (as it is currently in the cluster)
and the new object (passed as argument to apply). The original object is serialized and stored
in a special annotation (LabelApplied) every time an apply occurs - to be available for the
next apply. A similar mechanism is implemented in kubectl apply, which inspired wrangler apply.

When ignoreOriginal is true, the original object is read from its annotation only by the
getOriginalObject function. And, in turn, getOriginalObject is only called in 3 places:

  - in getOriginalBytes, but that never happens as it requires ignoreOriginal == false
  - in applyPatch (only if a reconciler is set)
  - in reconcileJob (a special reconciler case - the default reconciler for Job)

Thus, when ignoreOriginal is true and apart from reconciler cases,
  the original object annotation is never used.

This patch avoids computing and writing the original object annotation whenever ignoreOriginal
is true, as it won't be read and it is thus useless.

As an exception, the current behavior is maintained in reconciler cases.

Signed-off-by: Silvio Moioli <silvio@moioli.net>